css stuff, and invite rescind fix
