+++ Coop_Listing_Site/Coop_Listing_Site/Coop_Listing_Site.csproj
@@ -347,6 +347,8 @@
     <Content Include="Views\Department\Delete.cshtml" />
     <Content Include="Views\Department\Edit.cshtml" />
     <Content Include="Views\Department\Index.cshtml" />
+    <Content Include="Views\Student\Rescind.cshtml" />
+    <Content Include="Views\Coordinator\Rescind.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
