--- Coop_Listing_Site/Coop_Listing_Site/Views/ControlPanel/Rescind.cshtml
@@ -1,48 +1,0 @@
-﻿@model Coop_Listing_Site.Models.RegisterInvite
-
-@{
-    ViewBag.Title = "Rescind";
-    Layout = "~/Views/Shared/_CP_Layout.cshtml";
-}
-
-<h2>Rescind</h2>
-@if (ViewBag.Message != null)
-{
-    <h3>@ViewBag.Message</h3>
-}
-else
-{
-    <h3>Are you sure you want to rescind this invitation?</h3>
-}
-<div>
-    <h4>Invitation</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.Email)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.Email)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.UserType)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.UserType)
-        </dd>
-
-    </dl>
-
-    @using (Html.BeginForm())
-    {
-        @Html.AntiForgeryToken()
-
-        <div class="form-actions no-color">
-            <input type="submit" value="Delete" class="btn btn-default" /> |
-            @Html.ActionLink("Back to List", "InviteList")
-        </div>
-    }
-</div>
