+++ Coop_Listing_Site/Coop_Listing_Site/Views/Shared/_CP_Layout.cshtml
@@ -40,7 +40,6 @@
 </div>
 
     <div class="col-sm-9 col-sm-offset-3 col-md-10 col-md-offset-2 main"  id="main-wrapper">
-        <img src="~/Images/evening-sky-3.jpg" id="lcc-logo" />
             @RenderBody()
     </div>
